// src/app/modules/models/veicolo.model.ts
export class Veicolo{
    id:number;
    createdAt: number;
    updatedAt: number;
    userIns: string;
    acquisitionType: string;
    amount: number;
    brand: string;
    category: string;
    id_ram: string;
    licensePlate: string;
    model: string;
    type: string;
    adminState: string;
    adminUser: string;
    adminDateUpdate: string;
    adminNote: string;
    costoIstr: number;
    valoreContributo: number;
    pmiIstr: number;
    reteIstr: number;
    noteIstr: string;
}